requirements.txt
Django==4.2.7
PyMySQL==1.1.0
cryptography==41.0.7
openai==1.3.5
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7